DEV-1 WIP on push-container
The goal of this change is to reimplement push-container as part of
kube_deploy_tools to replace both Lance's push-docker-image.sh and
push-container.py from Arbor's workspace.

The role of push-container is to understand how to retag Docker images
in a way that agrees with generated Kubernetes manifests. For example, a
deploy manifest may specify:

containers:
  - image: <%= config["image_registry"] %>/mygreatapp:<%= config["tag"] %>

Suppose the source repo builds the Docker container 'mygreatapp'. The
role of push-container would be to retag it as:

REDACTED.dkr.ecr.us-west-2.amazonaws.com/mygreatapp:master_$GITSHA1

At the same time, render_deploys would set
image_registry = REDACTED.dkr.ecr.us-west-2.amazonaws.com
and
tag = master_$GITSHA1

since it is also part of kube_deploy_tools and has the same settings for
each target. Comparatively, push-docker-image.sh and push-container.py
have no awareness of the cluster configuration that is embedded into
kube_deploy_tools. Primarily, the goal of rewriting push-container is to
give it access to this shared source of truth.

50% of the work (the tagging) was already done as docker_deploy_tag.
This change would complete the work by promoting it back to a fully featured
push_container.

Here are the old scripts so you can see what they did:
https://git.REDACTED/arbortech/workspace/tree/master/scripts/push-container.py
https://git.REDACTED/ChefRepos/liveramp-jenkins-slave/blob/master/files/default/push-docker-image.sh