DEV-265 final fixes (#66)
In the prior commit, 7b10960b6f24e3f0e893d24b8a729b8a42970c0a, I added
functionality to make sure feature flags were added to the config hash.
Inadvertently, before we were omitting them by accident.

This makes up for the reviewless hotfix by adding testing around the
change. It ensures that non-string extra flags can be added to a
deploy.yml or cluster config, and they will show up in the final config
hash for evaluating at ERB time.

This required turning render_deploys_hook into a library. The bonus is
that we now get test coverage on it as long as the default hook is not
overridden, which it will never be within this project.

* tab -> space