pull_policy defaults to IfNotPresent
`IfNotPresent` is also the kubernetes default if not specified. We
probably want to use this default as well since with our tooling in
place (kdt + jenkins), we shouldn't have a real use case for repeatedly
updating a single tag (typically `latest`) and then relying on always
pulling images to update this tag on nodes. We should discourage any
remaining uses of the latest tag in this manner.