Sweep faster using batch commands (for AWS) (#139)
We use `aws ecr batch-delete-image`, but do not avail of its batch
functionality. Change the interface for the delete functionality of the
driver from delete_image to also include delete_images.

Each driver would choose whether to implement delete_images in terms of
delete_image (this is a default in the base driver) or whether to define
delete_images, then delete_image as a single case of delete_images.

For AWS, it's the latter, as the batch-delete-image functionality works
with one or the other. In practice, only delete_images is used by the
sweeper itself.