Docker: Drop a gcloud .docker/config.json inside the running container (#36)
Previously, KDT would self-write its own config.json, and this
was not hit in the docker image as a result.

However this feature was removed in
2389559c8edf5c046d65ba8ef275fafe0b92b116, because we wanted to let
downstream users choose how they authenticate with GCP. End users would
be responsible for calling `gcloud auth configure-docker`. Jenkins in
particular used `docker-credential-gcr` to avoid a dependency on
`gcloud` itself, and has its own config file.

IMO, this is still the correct move, so let's correct the docker image
by dropping the expected config.json into the home directory in the
entrypoint.