Support `libraries` parameter to extend a deploy.yaml with other ones (#45)
## Use case

Ideally, using KDT at LiveRamp should be as simple as including a file
which contains all LiveRamp specific defaults, such as the default
destination of Artifactory for all published files, a default gcp
artifact, the use of the gcp image registry, and any authentication
hookups which help use KDT with Jenkins.

## Solution

Define a `libraries` section of a deploy.yaml which permits downloading
an arbitrary number of local or gs:// files to extend this config with.
This way, we can define a gs://liveramp-kdt-configs/v1.yaml (or
osmething) which contains all of the above defaults.

Most people would use it this way:

```
version: 2
libraries:
  - gs://liveramp-kdt-configs/v1.yaml
```

Assuming v1 contained all the expected values, this should be a valid
KDT deploy.yaml and we would expect many apps to just use this verbatim.

### A note on precedence

The definition of extension is particular. When one deploy file
references a library, the in-memory model is extended with the contents
of the library file. Extension means different things for different
fields, but in general:

* Items in the library but not in the original file are merged into the
  original file.
* Items in the original file which conflict with the library will not be
  updated from the library.
* Library hooks will be executed first, in order, and your own hooks
  will be run at the end. This is because a library might be expected to
  prepare some stuff in the background to make execution of the default
  hooks possible or easier.

### Other cleanups

* This change removes a bunch of v1 compatibility. In v3 release of KDT,
   we intend to completely remove that support.